Add low-pass filtering.
This already existed in the common repository's mixer, but that was
a "brick-wall" filter, while this is a "first-order" filter. The
difference is that this gradually attenuates frequencies instead of
erasing them outright.

This should be much closer to an actual Mega Drive, as it is based on
a study of the console's actual circuitry. I'll still have to dig-out
my VA5 Mega Drive at some point just to be sure it's accurate.