Cache the VDP state pointer.
C and C++ have this dumb feature where it will re-read data from
memory in case a function call caused it to be modified, so we have
to write code like it's fucking assembly and manually cache data into
local variables to avoid needless reads. So much for high-level
programming.